Large scale rewrite
The primary motivations for the rewrite was to inhibit unnecessary
inlining by the compiler from the previous reification method. In
addition, this rewrite vastly reduces memory complexity of the
compilation and improves compiler throughput as well. Template
instantiations are aggressively avoided where possible, and work will
continue in this direction. As of this commit, testing is still missing
in PGA and a few operations are missing, but the nuts and bolts of the
framework are now here.